Put tests for reports in correct place and remove old stubs.
Some old stubs were kicking around that were implemented differently
than originally conceived. Also moves some code around to put tests for
the reports live in the correct directory.

Closes #230
